+++ src/Method/RestrictChatMemberMethod.php
@@ -8,6 +8,7 @@ use TgBotApi\BotApiBase\Method\Interfaces\RestrictMethodAliasInterface;
 use TgBotApi\BotApiBase\Method\Traits\ChatIdVariableTrait;
 use TgBotApi\BotApiBase\Method\Traits\FillFromArrayTrait;
 use TgBotApi\BotApiBase\Method\Traits\UserIdVariableTrait;
+use TgBotApi\BotApiBase\Type\ChatPermissionsType;
 
 /**
  * Class RestrictChatMemberMethod.
@@ -32,6 +33,8 @@ class RestrictChatMemberMethod implements RestrictMethodAliasInterface
     /**
      * Optional. Pass True, if the user can send text messages, contacts, locations and venues.
      *
+     * @deprecated
+     *
      * @var bool|null
      */
     public $canSendMessages;
@@ -39,6 +42,8 @@ class RestrictChatMemberMethod implements RestrictMethodAliasInterface
     /**
      * Optional. Pass True, if the user can send audios, documents, photos, videos, video notes and voice notes,
      * implies can_send_messages.
+     *
+     * @deprecated
      *
      * @var bool|null
      */
@@ -48,6 +53,8 @@ class RestrictChatMemberMethod implements RestrictMethodAliasInterface
      * Optional. Pass True, if the user can send animations, games, stickers and use inline bots,
      * implies can_send_media_messages.
      *
+     * @deprecated
+     *
      * @var bool|null
      */
     public $canSendOtherMessages;
@@ -55,24 +62,63 @@ class RestrictChatMemberMethod implements RestrictMethodAliasInterface
     /**
      * Optional. Pass True, if the user may add web page previews to their messages, implies can_send_media_messages.
      *
+     * @deprecated
+     *
      * @var bool|null
      */
     public $canAddWebPagePreviews;
 
     /**
-     * @param $chatId
+     * New user permissions.
+     *
+     * @var ChatPermissionsType
+     */
+    public $permissions;
+
+    /**
+     * @param            $chatId
      * @param int        $userId
      * @param array|null $data
      *
      * @throws \TgBotApi\BotApiBase\Exception\BadArgumentException
      *
      * @return RestrictChatMemberMethod
+     *
+     * @deprecated
+     * @see https://core.telegram.org/bots/api#july-29-2019
      */
-    public static function create($chatId, int $userId, array $data = null): RestrictChatMemberMethod
+    public static function createOld($chatId, int $userId, array $data = null): RestrictChatMemberMethod
     {
         $instance = new static();
         $instance->chatId = $chatId;
         $instance->userId = $userId;
+        if ($data) {
+            $instance->fill($data);
+        }
+
+        return $instance;
+    }
+
+    /**
+     * @param                     $chatId
+     * @param int                 $userId
+     * @param ChatPermissionsType $chatPermissionsType
+     * @param array|null          $data
+     *
+     * @throws \TgBotApi\BotApiBase\Exception\BadArgumentException
+     *
+     * @return RestrictChatMemberMethod
+     */
+    public static function create(
+        $chatId,
+        int $userId,
+        ChatPermissionsType $chatPermissionsType,
+        array $data = null
+    ): RestrictChatMemberMethod {
+        $instance = new static();
+        $instance->chatId = $chatId;
+        $instance->userId = $userId;
+        $instance->permissions = $chatPermissionsType;
         if ($data) {
             $instance->fill($data);
         }